hub & hubPath Refactoring
