+++ hub/src/main/java/com/team12/hub/domain/Hub.java
@@ -1,4 +1,4 @@
-package com.team12.hub.hub.domain;
+package com.team12.hub.domain;
 
 import com.team12.common.audit.AuditingEntity;
 import jakarta.persistence.*;
