+++ hub/src/main/java/com/team12/hub/dto/HubRequestDto.java
@@ -1,4 +1,4 @@
-package com.team12.hub.hub.dto;
+package com.team12.hub.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
