+++ hub/src/main/java/com/team12/hubPath/dto/HubPathRequestDto.java
@@ -1,4 +1,4 @@
-package com.team12.hub.hubPath.dto;
+package com.team12.hubPath.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
