+++ hub/src/main/java/com/team12/hubPath/dto/HubPathResponseDto.java
@@ -1,4 +1,4 @@
-package com.team12.hub.hubPath.dto;
+package com.team12.hubPath.dto;
 
 public class HubPathResponseDto {
